Cars data set from Kary Främling's PhD thesis.

Data collected from French car magazines in 1993.

data(CarsFramling)

Format

data.frame

Details

The columns are:

Car model

Price (french francs)

Maximum power (hp)

Type of transmission (automatic/manual, scores 2/4)

Traction (front-/rear-/four-wheel driven, scores 1/2/3)

Car size (m3)

Trunk size (liters)

Weight (kg)

Top speed (km/h)

Acceleration 1000m with zero start speed (s)

Fuel consumption (l/100km)

Aesthetic score (1, 2, 3, 4 or 5), subjective

Fabricant’s image (1, 2, 3, 4 or 5), subjective

Score for standard equipment (continuous from 0 to 83).

Preference value (continuous from 0 to 100).

References

Främling, Kary. Explaining Results of Neural Networks by Contextual Importance and Utility. In: Robert Andrews and Joachim Diederich (eds.), Rules and networks: Proceedings of the Rule Extraction from Trained Artificial Neural Networks Workshop, AISB'96 conference, 1-2 April 1996. Brighton, UK, 1996.

Främling, Kary. Modélisation et apprentissage des préférences par réseaux de neurones pour l'aide à la décision multicritère. PhD. thesis : Institut National de Sciences Appliquées de Lyon, Ecole Nationale Supérieure des Mines de Saint-Etienne, France, 1996. 209 p.

Examples

data(CarsFramling)