Barplot CIU explanation for specific instance

Create a barplot showing CI as the length of the bar and CU on color scale from red to green, via yellow, for the given inputs and the given output. First get a CIU object by calling ciu.new as e.g. ciu <- ciu.new(...), then call ciu.res <- ciu$barplot.ciu(...). "Usage" section is in "Details" section because Roxygen etc. don't support documentation of functions within functions.

Arguments

instance
Instance to explain. See explain.

ind.inputs
vector of indices for the inputs to be included in the plot. If NULL then all inputs will be included.

ind.output
Index of output to be explained.

in.min.max.limits
See explain.

n.samples
See explain.

neutral.CU
Indicates when the Contextual Utility is considered to be "negative". The default value of 0.5 seems quite logical for most cases.

show.input.values
Include input values after input labels or not. Default is TRUE.

concepts.to.explain
List of concepts to use in the plot, as defined by vocabulary provided as argument to ciu.new. If ind.inputs=NULL, then use concepts.to.explain instead. If both are NULL, then use all inputs.

target.concept
See explain.

target.ciu
See explain.

color.ramp.below.neutral
Color ramp function as returned by function colorRamp(). Default color ramp is from red3 to yellow.

color.ramp.above.neutral
Color ramp function as returned by function colorRamp(). Default colorramp is from yellow to darkgreen.

use.influence
Plot using "influence" rather than CIU, i.e. a LIME-like barplot. Default is FALSE.

sort
NULL, "CI" or "CU". No sorting by default, other options are sorting by CI or CU.

decreasing
Set to TRUE for decreasing sort.

main, xlab, xlim, ...
Usual plot parameters, possible to override the default ones provided here if needed.

Details

Usage

barplot.ciu(
  instance,
  ind.inputs=NULL,
  ind.output=1,
  in.min.max.limits=NULL,
  n.samples=100,
  neutral.CU=0.5,
  show.input.values=TRUE,
  concepts.to.explain=NULL,
  target.concept=NULL,
  target.ciu=NULL,
  color.ramp.below.neutral=NULL,
  color.ramp.above.neutral=NULL,
  sort=NULL,
  decreasing=FALSE,
  main=NULL,
  xlab=NULL,
  xlim=NULL,
  ...)

Value

"void", i.e. whatever happens to be result of last instruction.

See also

Author

Kary Främling