ciu.barplot

Create a barplot showing CI as the length of the bar and CU on color scale from red to green, via yellow, for the given inputs and the given output.

ciu.barplot(
  ciu,
  instance,
  ind.inputs = NULL,
  ind.output = 1,
  in.min.max.limits = NULL,
  n.samples = 100,
  neutral.CU = 0.5,
  show.input.values = TRUE,
  concepts.to.explain = NULL,
  target.concept = NULL,
  target.ciu = NULL,
  ciu.meta = NULL,
  color.ramp.below.neutral = NULL,
  color.ramp.above.neutral = NULL,
  use.influence = FALSE,
  influence.minmax = c(-1, 1),
  sort = NULL,
  decreasing = FALSE,
  main = NULL,
  xlab = NULL,
  xlim = NULL,
  ...
)

Arguments

ciu
ciu object as created with ciu function (not to be confused with CIU object as created by ciu.new).

instance
Input values for the instance to explain. Should be a data.frame even though a vector or matrix might work too if input names and other needed metadata can be deduced from the dataset or other parameters given to ciu.new.

ind.inputs
vector of indices for the inputs to be included in the plot. If NULL then all inputs will be included.

ind.output
Index of output to be explained.

in.min.max.limits
data.frame or matrix with one row per output and two columns, where the first column indicates the minimal value and the second column the maximal value for that output. ONLY NEEDED HERE IF not given as parameter to ciu.new or if the limits are different for this specific instance than the default ones.

n.samples
How many instances to generate for estimating CI and CU. For inputs of type factor, all possible combinations of input values are generated, so this parameter only influences how many instances are (at least) generated for continuous-valued inputs.

neutral.CU
Indicates when the Contextual Utility is considered to be "negative". The default value of 0.5 seems quite logical for most cases.

show.input.values
Include input values after input labels or not. Default is TRUE.

concepts.to.explain
List of concepts to use in the plot, as defined by vocabulary provided as argument to ciu.new. If ind.inputs=NULL, then use concepts.to.explain instead. If both are NULL, then use all inputs.

target.concept
If provided, then calculate CIU of inputs ind.inputs.to.explain relative to the given concept rather than relative to the actual output(s). ind.inputs.to.explain should normally be a subset (or all) of the inputs that target.concept consists of, even though that not required by the CIU calculation. If a target.ciu is provided, then the target.concept doesn't have to be included in the vocabulary gives as parameter to ciu.new (at least for the moment).

target.ciu
ciu.result object previously calculated for target.concept. If a target.concept is provided but target.ciu=NULL, then target.ciu is estimated by a call to ciu.explain with the n.samples value given as a parameter to this call. It may be useful to provide target.ciu if it should be estimated using some other (typically greater) value for n.samples than the default one, or if it has already been calculated for some reason.

ciu.meta
If given, then use existing ciu.meta.result rather than calling ciu.meta.explain.

color.ramp.below.neutral
Color ramp function as returned by function colorRamp(). Default color ramp is from red3 to yellow.

color.ramp.above.neutral
Color ramp function as returned by function colorRamp(). Default colorramp is from yellow to darkgreen.

use.influence
Plot using "influence" rather than CIU, i.e. a LIME-like barplot. Default is FALSE.

influence.minmax
Range to use for influence values.

sort
NULL, "CI" or "CU".

decreasing
Set to TRUE for decreasing sort.

main
Text to use as main title.

xlab
Label for x-axis.

xlim
Minimal and maximal values for x-axis.

...
See base::plot.

Value

"void", i.e. whatever happens to be result of last instruction.

See also

Author

Kary Främling