ciu.plot

Function for plotting out the effect of changing values of one input on one output

ciu.plot(
  ciu,
  instance,
  ind.input,
  ind.output,
  in.min.max.limits = NULL,
  n.points = 40,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  ...
)

Arguments

ciu

ciu object as created with ciu function (not to be confused with CIU object as created by ciu.new).

instance

Input values for the instance to explain. Should be a data.frame even though a vector or matrix might work too if input names and other needed metadata can be deduced from the dataset or other parameters given to ciu.new.

ind.input

Index of input feature to plot.

ind.output

Index of output to plot.

in.min.max.limits

data.frame or matrix with one row per output and two columns, where the first column indicates the minimal value and the second column the maximal value for that output. ONLY NEEDED HERE IF not given as parameter to ciu.new or if the limits are different for this specific instance than the default ones.

n.points

How many x,y pairs will be calculated, equally spaced over in.min.max.limits.

main

Text to use as main title.

xlab

Label for x-axis.

ylab

Label for y-axis.

ylim

Minimal and maximal values for y-axis.

...

See base::plot.

Value

"void", or whatever happens to be result of last instruction.

See also

Author

Kary Främling