Give textual CIU explanation

Provide textual CIU explanations as those used in Kary Främling's PhD thesis.

ciu.textual(
  ciu,
  instance = NULL,
  ind.inputs = NULL,
  ind.output = 1,
  in.min.max.limits = NULL,
  n.samples = 100,
  neutral.CU = 0.5,
  show.input.values = TRUE,
  concepts.to.explain = NULL,
  target.concept = NULL,
  target.ciu = NULL,
  ciu.meta = NULL,
  sort = "CI",
  n.features = NULL,
  use.text.effects = FALSE,
  CI.voc = data.frame(limits = c(0.2, 0.4, 0.6, 0.8, 1), texts = c("not important",
    "slightly important", "important", "very important", "extremely important")),
  CU.voc = data.frame(limits = c(0.2, 0.4, 0.6, 0.8, 1), texts = c("very bad", "bad",
    "average", "good", "very good"))
)

Arguments

ciu
ciu object as created with ciu function (not to be confused with CIU object as created by ciu.new).

instance
Input values for the instance to explain. Should be a data.frame even though a vector or matrix might work too if input names and other needed metadata can be deduced from the dataset or other parameters given to ciu.new.

ind.inputs
Indices of input features to explain (the set i in CIU formulae)

ind.output
Index of output to be explained.

in.min.max.limits
data.frame or matrix with one row per output and two columns, where the first column indicates the minimal value and the second column the maximal value for that output. ONLY NEEDED HERE IF not given as parameter to ciu.new or if the limits are different for this specific instance than the default ones.

n.samples
How many instances to generate for estimating CI and CU. For inputs of type factor, all possible combinations of input values are generated, so this parameter only influences how many instances are (at least) generated for continuous-valued inputs.

neutral.CU
Indicates when the Contextual Utility is considered to be "negative". The default value of 0.5 seems quite logical for most cases.

show.input.values
Include input values after input labels or not. Default is TRUE.

concepts.to.explain
List of input feature concepts to explain, as defined by vocabulary provided as argument to ciu.new. If ind.inputs=NULL, then use concepts.to.explain instead. If both are NULL, then use all inputs.

target.concept
If provided, then calculate CIU of inputs ind.inputs.to.explain relative to the given concept rather than relative to the actual output(s). ind.inputs.to.explain should normally be a subset (or all) of the inputs that target.concept consists of, even though that not required by the CIU calculation. If a target.ciu is provided, then the target.concept doesn't have to be included in the vocabulary gives as parameter to ciu.new (at least for the moment).

target.ciu
ciu.result object previously calculated for target.concept. If a target.concept is provided but target.ciu=NULL, then target.ciu is estimated by a call to ciu.explain with the n.samples value given as a parameter to this call. It may be useful to provide target.ciu if it should be estimated using some other (typically greater) value for n.samples than the default one, or if it has already been calculated for some reason.

ciu.meta
If given, then use existing ciu.meta.result rather than calling ciu.meta.explain.

sort
NULL, "CI" or "CU".

n.features
Maximal number of features to include in explanation.

use.text.effects
Add bold/italics/colors effects?

CI.voc
Limits and texts to use for CI values.

CU.voc
Limits and texts to use for CU values.

Value

Text string with explanation.

Examples

# Explaining the classification of an Iris instance with lda model.
# We use a versicolor (instance 100).
library(MASS)
test.ind <- 100
iris_test <- iris[test.ind, 1:4]
iris_train <- iris[-test.ind, 1:4]
iris_lab <- iris[[5]][-test.ind]
model <- lda(iris_train, iris_lab)

# Create CIU object
ciu <- ciu.new(model, Species~., iris)

# Give textual explanation. Use 'cat' for getting newlines to work.
cat(ciu.textual(ciu, iris_test, ind.output = 2))

#> The value of output 'versicolor' for instance '100' is 1, which is very good (CU=1).
#> Feature 'Petal.Length' is extremely important (CI=1) and value '4.1' is very good (CU=1).
#> Feature 'Petal.Width' is extremely important (CI=1) and value '1.3' is very good (CU=1).
#> Feature 'Sepal.Length' is not important (CI=0.007) and value '5.7' is very good (CU=0.989).
#> Feature 'Sepal.Width' is not important (CI=0.001) and value '2.8' is very good (CU=0.931).

cat(ciu.textual(ciu, iris_test, ind.output = 2, n.features = 2))

#> The value of output 'versicolor' for instance '100' is 1, which is very good (CU=1).
#> Feature 'Petal.Length' is extremely important (CI=1) and value '4.1' is very good (CU=1).
#> Feature 'Petal.Width' is extremely important (CI=1) and value '1.3' is very good (CU=1).

# Boston housing, GBM model.
library(caret)
kfoldcv <- trainControl(method="cv", number=10)
gbm <- train(medv ~ ., Boston, method="gbm", trControl=kfoldcv)

#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       76.2591             nan     0.1000    7.5628
#>      2       69.5486             nan     0.1000    6.2330
#>      3       63.7750             nan     0.1000    6.0730
#>      4       58.8662             nan     0.1000    4.6601
#>      5       55.1082             nan     0.1000    3.4711
#>      6       51.4823             nan     0.1000    3.5004
#>      7       47.8226             nan     0.1000    3.3589
#>      8       44.9082             nan     0.1000    2.0835
#>      9       42.3230             nan     0.1000    2.7820
#>     10       39.4836             nan     0.1000    2.8217
#>     20       24.8577             nan     0.1000    0.7921
#>     40       16.4523             nan     0.1000   -0.0065
#>     60       13.6742             nan     0.1000   -0.0018
#>     80       12.2185             nan     0.1000    0.0287
#>    100       11.3203             nan     0.1000   -0.0027
#>    120       10.8571             nan     0.1000   -0.0086
#>    140       10.3810             nan     0.1000   -0.0423
#>    150       10.2197             nan     0.1000   -0.1238
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       73.8769             nan     0.1000   10.0242
#>      2       65.5702             nan     0.1000    6.9999
#>      3       59.2885             nan     0.1000    7.0203
#>      4       52.5204             nan     0.1000    5.7978
#>      5       47.1798             nan     0.1000    5.4333
#>      6       42.5385             nan     0.1000    4.2663
#>      7       39.4964             nan     0.1000    3.4309
#>      8       35.9388             nan     0.1000    2.9293
#>      9       33.2519             nan     0.1000    1.7745
#>     10       30.6738             nan     0.1000    1.6901
#>     20       17.7835             nan     0.1000    0.1399
#>     40       11.5822             nan     0.1000    0.0153
#>     60        9.5199             nan     0.1000   -0.0551
#>     80        8.6156             nan     0.1000   -0.0098
#>    100        7.7405             nan     0.1000    0.0196
#>    120        7.1387             nan     0.1000   -0.0939
#>    140        6.5823             nan     0.1000   -0.0555
#>    150        6.3415             nan     0.1000    0.0042
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       73.1077             nan     0.1000   11.5329
#>      2       64.2878             nan     0.1000    8.0911
#>      3       57.0295             nan     0.1000    7.1097
#>      4       50.3848             nan     0.1000    6.5817
#>      5       44.9702             nan     0.1000    5.9034
#>      6       40.3468             nan     0.1000    4.5810
#>      7       36.6667             nan     0.1000    3.1286
#>      8       33.3789             nan     0.1000    3.3781
#>      9       29.8627             nan     0.1000    2.6910
#>     10       27.4359             nan     0.1000    2.0225
#>     20       14.8274             nan     0.1000    0.5992
#>     40        9.0036             nan     0.1000   -0.0290
#>     60        7.5580             nan     0.1000   -0.0127
#>     80        6.4790             nan     0.1000   -0.0786
#>    100        5.6551             nan     0.1000   -0.0350
#>    120        5.1151             nan     0.1000   -0.0408
#>    140        4.5724             nan     0.1000   -0.0125
#>    150        4.3880             nan     0.1000   -0.0398
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       76.1872             nan     0.1000    6.3247
#>      2       69.9267             nan     0.1000    6.2674
#>      3       64.1521             nan     0.1000    4.4469
#>      4       59.5040             nan     0.1000    4.4372
#>      5       54.5452             nan     0.1000    3.9037
#>      6       50.8463             nan     0.1000    3.6234
#>      7       47.2157             nan     0.1000    3.2993
#>      8       43.9362             nan     0.1000    2.7303
#>      9       41.2905             nan     0.1000    2.4109
#>     10       38.8103             nan     0.1000    2.7804
#>     20       24.3098             nan     0.1000    0.9876
#>     40       15.8434             nan     0.1000    0.1520
#>     60       13.3665             nan     0.1000    0.0243
#>     80       12.1780             nan     0.1000    0.0182
#>    100       11.3603             nan     0.1000   -0.0109
#>    120       10.8155             nan     0.1000    0.0024
#>    140       10.4011             nan     0.1000   -0.0699
#>    150       10.2513             nan     0.1000    0.0077
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       73.0765             nan     0.1000    9.3541
#>      2       64.5685             nan     0.1000    7.2486
#>      3       58.3003             nan     0.1000    6.0589
#>      4       52.4381             nan     0.1000    4.7712
#>      5       46.2977             nan     0.1000    5.7942
#>      6       41.6000             nan     0.1000    4.3162
#>      7       37.9022             nan     0.1000    2.0698
#>      8       34.6525             nan     0.1000    2.3401
#>      9       31.7369             nan     0.1000    2.1746
#>     10       29.4499             nan     0.1000    1.8093
#>     20       16.8102             nan     0.1000    0.4176
#>     40       11.5105             nan     0.1000    0.0366
#>     60        9.6217             nan     0.1000   -0.0031
#>     80        8.7167             nan     0.1000   -0.0060
#>    100        7.7405             nan     0.1000   -0.0509
#>    120        7.2336             nan     0.1000   -0.0414
#>    140        6.7096             nan     0.1000   -0.0385
#>    150        6.4323             nan     0.1000    0.0171
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       72.2837             nan     0.1000    8.3220
#>      2       63.0160             nan     0.1000    7.2741
#>      3       54.8847             nan     0.1000    7.3763
#>      4       49.5994             nan     0.1000    6.2478
#>      5       44.2620             nan     0.1000    5.6443
#>      6       39.6348             nan     0.1000    4.0689
#>      7       35.2465             nan     0.1000    2.8308
#>      8       31.4876             nan     0.1000    3.5556
#>      9       28.6040             nan     0.1000    1.8128
#>     10       26.1117             nan     0.1000    2.1106
#>     20       14.3194             nan     0.1000    0.0914
#>     40        9.1207             nan     0.1000    0.0817
#>     60        7.3512             nan     0.1000   -0.0257
#>     80        6.3737             nan     0.1000   -0.0386
#>    100        5.6078             nan     0.1000   -0.0174
#>    120        5.0079             nan     0.1000   -0.0460
#>    140        4.5332             nan     0.1000   -0.0098
#>    150        4.3255             nan     0.1000   -0.0616
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       79.6155             nan     0.1000    6.9223
#>      2       72.1806             nan     0.1000    7.2957
#>      3       67.0551             nan     0.1000    4.9196
#>      4       61.8029             nan     0.1000    4.4664
#>      5       57.5844             nan     0.1000    4.3396
#>      6       54.0067             nan     0.1000    3.2147
#>      7       50.0329             nan     0.1000    3.8414
#>      8       46.9559             nan     0.1000    3.1844
#>      9       43.9095             nan     0.1000    2.6830
#>     10       41.1705             nan     0.1000    2.0684
#>     20       25.9322             nan     0.1000    0.6684
#>     40       17.3553             nan     0.1000    0.2105
#>     60       14.2829             nan     0.1000    0.0409
#>     80       12.8218             nan     0.1000   -0.2870
#>    100       11.9683             nan     0.1000   -0.0715
#>    120       11.2952             nan     0.1000    0.0092
#>    140       10.7616             nan     0.1000   -0.0494
#>    150       10.5988             nan     0.1000   -0.0923
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       76.6967             nan     0.1000    9.0039
#>      2       68.2144             nan     0.1000    8.1178
#>      3       60.3659             nan     0.1000    6.5987
#>      4       54.5435             nan     0.1000    5.1737
#>      5       48.6338             nan     0.1000    5.6894
#>      6       44.0215             nan     0.1000    4.1405
#>      7       40.8054             nan     0.1000    2.6324
#>      8       37.3064             nan     0.1000    2.7578
#>      9       34.1552             nan     0.1000    2.7543
#>     10       31.9116             nan     0.1000    2.0974
#>     20       18.5822             nan     0.1000    0.4681
#>     40       12.1438             nan     0.1000    0.0432
#>     60       10.1601             nan     0.1000   -0.0390
#>     80        8.8939             nan     0.1000   -0.1370
#>    100        8.2387             nan     0.1000   -0.0608
#>    120        7.7054             nan     0.1000   -0.0726
#>    140        7.2439             nan     0.1000   -0.0404
#>    150        7.0510             nan     0.1000   -0.0697
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       75.6802             nan     0.1000   11.4719
#>      2       66.4589             nan     0.1000    8.8891
#>      3       58.8244             nan     0.1000    7.7539
#>      4       51.8349             nan     0.1000    6.0274
#>      5       46.6558             nan     0.1000    5.7745
#>      6       41.7156             nan     0.1000    5.0929
#>      7       38.2256             nan     0.1000    2.4589
#>      8       34.5399             nan     0.1000    3.4505
#>      9       32.0480             nan     0.1000    2.8146
#>     10       28.9975             nan     0.1000    2.5740
#>     20       16.1416             nan     0.1000    0.4903
#>     40        9.7386             nan     0.1000   -0.0445
#>     60        8.0029             nan     0.1000   -0.0218
#>     80        7.0036             nan     0.1000   -0.1056
#>    100        6.1797             nan     0.1000   -0.0161
#>    120        5.4757             nan     0.1000   -0.0386
#>    140        5.0386             nan     0.1000   -0.0401
#>    150        4.7736             nan     0.1000   -0.0212
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       75.1330             nan     0.1000    6.0746
#>      2       69.3701             nan     0.1000    5.2076
#>      3       63.5167             nan     0.1000    6.0198
#>      4       59.1004             nan     0.1000    4.2871
#>      5       54.6619             nan     0.1000    4.2297
#>      6       50.4811             nan     0.1000    4.2585
#>      7       47.2489             nan     0.1000    3.2541
#>      8       43.9265             nan     0.1000    2.8802
#>      9       41.0957             nan     0.1000    2.6132
#>     10       38.8106             nan     0.1000    1.9113
#>     20       23.7644             nan     0.1000    0.4750
#>     40       15.5956             nan     0.1000    0.1234
#>     60       12.7476             nan     0.1000    0.0347
#>     80       11.2856             nan     0.1000    0.0151
#>    100       10.4858             nan     0.1000   -0.0805
#>    120        9.8944             nan     0.1000   -0.0105
#>    140        9.5572             nan     0.1000   -0.1032
#>    150        9.3365             nan     0.1000   -0.0162
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       72.2869             nan     0.1000   10.6663
#>      2       64.4859             nan     0.1000    7.5401
#>      3       57.5844             nan     0.1000    6.1275
#>      4       51.4161             nan     0.1000    5.3455
#>      5       45.8887             nan     0.1000    4.6521
#>      6       41.7772             nan     0.1000    3.2838
#>      7       37.9263             nan     0.1000    3.9922
#>      8       34.4718             nan     0.1000    2.4420
#>      9       31.4920             nan     0.1000    2.4696
#>     10       28.7406             nan     0.1000    2.6477
#>     20       16.4876             nan     0.1000    0.4207
#>     40       10.6197             nan     0.1000    0.0470
#>     60        8.6790             nan     0.1000    0.0065
#>     80        7.8222             nan     0.1000   -0.0672
#>    100        7.0828             nan     0.1000   -0.0175
#>    120        6.5830             nan     0.1000   -0.0385
#>    140        6.0782             nan     0.1000   -0.0124
#>    150        5.8433             nan     0.1000   -0.0413
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       71.4507             nan     0.1000   11.5641
#>      2       62.3852             nan     0.1000    9.2463
#>      3       55.2373             nan     0.1000    6.3463
#>      4       48.7719             nan     0.1000    5.2253
#>      5       43.5217             nan     0.1000    4.2678
#>      6       39.5485             nan     0.1000    3.3408
#>      7       35.7484             nan     0.1000    3.4178
#>      8       32.3498             nan     0.1000    4.0377
#>      9       29.2096             nan     0.1000    2.8366
#>     10       26.7895             nan     0.1000    2.0385
#>     20       13.9888             nan     0.1000    0.2999
#>     40        8.3789             nan     0.1000    0.0221
#>     60        6.9385             nan     0.1000   -0.0682
#>     80        6.0881             nan     0.1000   -0.0904
#>    100        5.3537             nan     0.1000   -0.0227
#>    120        4.8221             nan     0.1000   -0.0731
#>    140        4.3172             nan     0.1000    0.0062
#>    150        4.1316             nan     0.1000   -0.0217
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       77.8598             nan     0.1000    6.3020
#>      2       71.3340             nan     0.1000    6.4438
#>      3       65.8758             nan     0.1000    4.0920
#>      4       61.0456             nan     0.1000    5.4616
#>      5       56.5681             nan     0.1000    4.3662
#>      6       52.2885             nan     0.1000    3.8628
#>      7       48.9048             nan     0.1000    2.7483
#>      8       46.0225             nan     0.1000    2.8702
#>      9       43.0957             nan     0.1000    2.1569
#>     10       40.4054             nan     0.1000    2.3094
#>     20       25.6381             nan     0.1000    0.8389
#>     40       16.5988             nan     0.1000    0.1298
#>     60       13.7568             nan     0.1000    0.0029
#>     80       12.4240             nan     0.1000   -0.0434
#>    100       11.2982             nan     0.1000   -0.0248
#>    120       10.7245             nan     0.1000   -0.1600
#>    140       10.4308             nan     0.1000   -0.0012
#>    150       10.1985             nan     0.1000   -0.1470
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       74.8214             nan     0.1000   10.1241
#>      2       65.8997             nan     0.1000    8.1566
#>      3       60.5201             nan     0.1000    5.0386
#>      4       53.6519             nan     0.1000    6.3239
#>      5       48.2712             nan     0.1000    5.5084
#>      6       43.5815             nan     0.1000    4.7071
#>      7       39.7245             nan     0.1000    3.5437
#>      8       35.7608             nan     0.1000    4.0861
#>      9       32.6605             nan     0.1000    2.7354
#>     10       30.1618             nan     0.1000    2.5436
#>     20       17.9728             nan     0.1000    0.6741
#>     40       11.5605             nan     0.1000   -0.0332
#>     60        9.4110             nan     0.1000    0.0167
#>     80        8.4219             nan     0.1000   -0.0184
#>    100        7.6631             nan     0.1000   -0.0170
#>    120        7.0052             nan     0.1000   -0.0469
#>    140        6.5459             nan     0.1000   -0.0758
#>    150        6.2873             nan     0.1000   -0.0151
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       74.3001             nan     0.1000    9.9769
#>      2       65.2392             nan     0.1000    8.3359
#>      3       57.2327             nan     0.1000    8.2877
#>      4       50.7483             nan     0.1000    6.6220
#>      5       45.4182             nan     0.1000    4.7193
#>      6       41.0420             nan     0.1000    3.8432
#>      7       37.6497             nan     0.1000    3.1963
#>      8       33.8473             nan     0.1000    2.7631
#>      9       30.7353             nan     0.1000    2.4887
#>     10       27.9822             nan     0.1000    2.5057
#>     20       15.2926             nan     0.1000    0.2786
#>     40        9.8431             nan     0.1000   -0.1105
#>     60        8.0668             nan     0.1000    0.0021
#>     80        6.9418             nan     0.1000   -0.0307
#>    100        6.0741             nan     0.1000    0.0017
#>    120        5.4658             nan     0.1000   -0.0541
#>    140        4.9504             nan     0.1000   -0.0171
#>    150        4.7240             nan     0.1000   -0.0366
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       76.2163             nan     0.1000    6.5431
#>      2       69.9999             nan     0.1000    5.9542
#>      3       63.9317             nan     0.1000    6.1098
#>      4       59.4516             nan     0.1000    4.8184
#>      5       55.2967             nan     0.1000    3.9735
#>      6       51.1552             nan     0.1000    3.8673
#>      7       47.4731             nan     0.1000    3.6901
#>      8       44.3824             nan     0.1000    2.6015
#>      9       41.5764             nan     0.1000    2.7742
#>     10       39.0908             nan     0.1000    2.0047
#>     20       23.7582             nan     0.1000    1.0314
#>     40       14.9078             nan     0.1000    0.0792
#>     60       12.5274             nan     0.1000   -0.0690
#>     80       11.3530             nan     0.1000    0.0530
#>    100       10.6020             nan     0.1000   -0.0893
#>    120       10.1767             nan     0.1000   -0.0740
#>    140        9.8415             nan     0.1000   -0.0285
#>    150        9.5789             nan     0.1000   -0.0040
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       73.3557             nan     0.1000    9.7882
#>      2       65.6627             nan     0.1000    6.7281
#>      3       58.6797             nan     0.1000    5.7559
#>      4       53.4418             nan     0.1000    4.5726
#>      5       47.9170             nan     0.1000    5.5334
#>      6       43.2687             nan     0.1000    4.6459
#>      7       39.2942             nan     0.1000    3.5158
#>      8       35.6056             nan     0.1000    3.1959
#>      9       32.9212             nan     0.1000    2.3664
#>     10       30.5463             nan     0.1000    1.9963
#>     20       17.6280             nan     0.1000    0.4758
#>     40       11.3740             nan     0.1000    0.1056
#>     60        9.3932             nan     0.1000    0.0141
#>     80        8.2093             nan     0.1000   -0.0101
#>    100        7.5175             nan     0.1000    0.0049
#>    120        6.9336             nan     0.1000   -0.0293
#>    140        6.4417             nan     0.1000   -0.0174
#>    150        6.2273             nan     0.1000   -0.0490
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       72.9431             nan     0.1000    8.0809
#>      2       62.9074             nan     0.1000    8.0159
#>      3       55.9513             nan     0.1000    5.9871
#>      4       48.9463             nan     0.1000    6.4561
#>      5       43.7678             nan     0.1000    5.1497
#>      6       39.0264             nan     0.1000    3.9649
#>      7       34.9353             nan     0.1000    3.7195
#>      8       31.4044             nan     0.1000    2.7247
#>      9       28.2016             nan     0.1000    2.6881
#>     10       25.7385             nan     0.1000    2.4567
#>     20       13.8983             nan     0.1000    0.2427
#>     40        9.0307             nan     0.1000   -0.0716
#>     60        7.4110             nan     0.1000   -0.0296
#>     80        6.3676             nan     0.1000   -0.0551
#>    100        5.6426             nan     0.1000   -0.0337
#>    120        5.1045             nan     0.1000   -0.0208
#>    140        4.6262             nan     0.1000   -0.0174
#>    150        4.4171             nan     0.1000   -0.0260
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       77.9481             nan     0.1000    6.1480
#>      2       70.9458             nan     0.1000    6.4085
#>      3       65.1518             nan     0.1000    4.4513
#>      4       60.3635             nan     0.1000    3.7718
#>      5       56.0187             nan     0.1000    3.9127
#>      6       52.4185             nan     0.1000    3.8521
#>      7       49.1744             nan     0.1000    3.0732
#>      8       46.2544             nan     0.1000    2.9847
#>      9       43.4127             nan     0.1000    2.9755
#>     10       40.3921             nan     0.1000    3.0632
#>     20       25.3792             nan     0.1000    0.7600
#>     40       16.4319             nan     0.1000    0.1001
#>     60       13.7149             nan     0.1000   -0.0814
#>     80       12.2087             nan     0.1000   -0.0413
#>    100       11.2515             nan     0.1000    0.0353
#>    120       10.5787             nan     0.1000   -0.0010
#>    140       10.1709             nan     0.1000    0.0006
#>    150        9.9831             nan     0.1000   -0.0584
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       75.3423             nan     0.1000    9.9765
#>      2       67.5008             nan     0.1000    7.5213
#>      3       59.4751             nan     0.1000    7.4081
#>      4       53.4611             nan     0.1000    5.7839
#>      5       48.2873             nan     0.1000    5.1276
#>      6       44.1582             nan     0.1000    4.2414
#>      7       40.6281             nan     0.1000    2.5129
#>      8       37.3938             nan     0.1000    3.3864
#>      9       34.3938             nan     0.1000    2.6495
#>     10       32.0514             nan     0.1000    2.4105
#>     20       19.2704             nan     0.1000    0.5212
#>     40       12.5858             nan     0.1000    0.1844
#>     60       10.4727             nan     0.1000    0.0635
#>     80        9.3267             nan     0.1000   -0.0579
#>    100        8.6108             nan     0.1000   -0.0386
#>    120        7.9239             nan     0.1000   -0.0123
#>    140        7.3076             nan     0.1000   -0.0781
#>    150        7.1011             nan     0.1000   -0.0273
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       74.2272             nan     0.1000   13.2233
#>      2       64.3309             nan     0.1000    8.7129
#>      3       56.4120             nan     0.1000    7.0939
#>      4       50.5270             nan     0.1000    5.0343
#>      5       45.0340             nan     0.1000    4.3336
#>      6       40.3526             nan     0.1000    4.4005
#>      7       36.2121             nan     0.1000    2.5233
#>      8       32.8776             nan     0.1000    3.3135
#>      9       29.8470             nan     0.1000    2.3721
#>     10       27.4908             nan     0.1000    2.3496
#>     20       15.2166             nan     0.1000    0.7017
#>     40        9.5176             nan     0.1000    0.0583
#>     60        7.5932             nan     0.1000    0.0326
#>     80        6.4884             nan     0.1000   -0.0034
#>    100        5.7419             nan     0.1000   -0.0719
#>    120        5.1336             nan     0.1000   -0.0357
#>    140        4.6896             nan     0.1000   -0.0374
#>    150        4.4797             nan     0.1000   -0.0403
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       79.5201             nan     0.1000    6.1178
#>      2       73.8021             nan     0.1000    4.9228
#>      3       67.2739             nan     0.1000    5.9020
#>      4       62.0203             nan     0.1000    4.8813
#>      5       57.6458             nan     0.1000    4.4789
#>      6       53.6036             nan     0.1000    3.9989
#>      7       50.0986             nan     0.1000    3.5902
#>      8       47.1277             nan     0.1000    2.7419
#>      9       44.1321             nan     0.1000    2.0824
#>     10       41.4796             nan     0.1000    1.9392
#>     20       26.4875             nan     0.1000    0.3516
#>     40       16.9305             nan     0.1000    0.1000
#>     60       13.8884             nan     0.1000   -0.0377
#>     80       12.3996             nan     0.1000    0.0204
#>    100       11.5751             nan     0.1000   -0.1362
#>    120       11.0358             nan     0.1000   -0.0212
#>    140       10.5592             nan     0.1000   -0.0117
#>    150       10.3644             nan     0.1000   -0.0252
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       76.0780             nan     0.1000    9.4212
#>      2       67.5567             nan     0.1000    8.3758
#>      3       60.6994             nan     0.1000    5.9420
#>      4       54.4181             nan     0.1000    5.9121
#>      5       48.5764             nan     0.1000    5.6993
#>      6       44.0209             nan     0.1000    3.9681
#>      7       39.8496             nan     0.1000    3.1359
#>      8       36.6871             nan     0.1000    2.7862
#>      9       34.0755             nan     0.1000    2.7044
#>     10       31.7755             nan     0.1000    1.6097
#>     20       18.3297             nan     0.1000    0.8180
#>     40       11.9479             nan     0.1000   -0.0002
#>     60        9.8488             nan     0.1000   -0.1252
#>     80        8.8365             nan     0.1000   -0.0864
#>    100        8.1800             nan     0.1000   -0.0339
#>    120        7.4160             nan     0.1000   -0.0182
#>    140        6.9970             nan     0.1000   -0.0256
#>    150        6.7862             nan     0.1000   -0.0340
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       75.5945             nan     0.1000   10.9400
#>      2       66.0086             nan     0.1000   10.0679
#>      3       57.6400             nan     0.1000    7.1766
#>      4       50.8961             nan     0.1000    5.7406
#>      5       44.8197             nan     0.1000    5.2555
#>      6       40.3577             nan     0.1000    4.2006
#>      7       36.4179             nan     0.1000    3.4066
#>      8       33.0099             nan     0.1000    3.4015
#>      9       29.8016             nan     0.1000    2.7009
#>     10       27.4301             nan     0.1000    1.9642
#>     20       15.0409             nan     0.1000    0.5245
#>     40        9.7391             nan     0.1000    0.0474
#>     60        8.2685             nan     0.1000   -0.0689
#>     80        7.0055             nan     0.1000   -0.0538
#>    100        6.3607             nan     0.1000   -0.0947
#>    120        5.6510             nan     0.1000   -0.0202
#>    140        5.0824             nan     0.1000   -0.0344
#>    150        4.8385             nan     0.1000   -0.0381
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       75.5465             nan     0.1000    5.4788
#>      2       69.1603             nan     0.1000    6.0778
#>      3       63.2246             nan     0.1000    4.8806
#>      4       59.1449             nan     0.1000    4.1115
#>      5       54.5199             nan     0.1000    4.1485
#>      6       50.8169             nan     0.1000    3.5699
#>      7       47.7311             nan     0.1000    3.3231
#>      8       44.9653             nan     0.1000    2.7207
#>      9       42.2781             nan     0.1000    2.1379
#>     10       39.5394             nan     0.1000    2.2089
#>     20       24.3043             nan     0.1000    0.7964
#>     40       15.3710             nan     0.1000    0.1405
#>     60       12.9132             nan     0.1000    0.0463
#>     80       11.4444             nan     0.1000   -0.0047
#>    100       10.6942             nan     0.1000   -0.0643
#>    120       10.1385             nan     0.1000    0.0050
#>    140        9.7149             nan     0.1000   -0.0782
#>    150        9.5162             nan     0.1000   -0.0967
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       73.3066             nan     0.1000    9.8045
#>      2       65.6331             nan     0.1000    8.9520
#>      3       58.7864             nan     0.1000    7.2757
#>      4       52.8351             nan     0.1000    5.4469
#>      5       47.2156             nan     0.1000    4.9209
#>      6       42.7812             nan     0.1000    4.4635
#>      7       38.4517             nan     0.1000    4.3259
#>      8       35.3579             nan     0.1000    2.8338
#>      9       32.3378             nan     0.1000    2.0587
#>     10       29.6520             nan     0.1000    2.2420
#>     20       17.1456             nan     0.1000    0.5849
#>     40       11.3369             nan     0.1000   -0.0087
#>     60        9.5412             nan     0.1000   -0.1305
#>     80        8.4329             nan     0.1000   -0.1078
#>    100        7.7784             nan     0.1000   -0.0213
#>    120        6.9522             nan     0.1000   -0.0408
#>    140        6.3818             nan     0.1000   -0.0241
#>    150        6.1261             nan     0.1000   -0.0264
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       71.0331             nan     0.1000   11.9464
#>      2       62.0485             nan     0.1000    8.1611
#>      3       54.3011             nan     0.1000    7.7131
#>      4       48.0800             nan     0.1000    5.4041
#>      5       43.0980             nan     0.1000    4.4833
#>      6       38.7356             nan     0.1000    3.4513
#>      7       35.0675             nan     0.1000    3.1463
#>      8       31.9426             nan     0.1000    2.5077
#>      9       28.5988             nan     0.1000    2.7796
#>     10       26.2249             nan     0.1000    2.5077
#>     20       14.4424             nan     0.1000    0.5029
#>     40        9.1252             nan     0.1000   -0.0565
#>     60        7.7090             nan     0.1000   -0.0627
#>     80        6.5431             nan     0.1000   -0.0261
#>    100        5.9327             nan     0.1000   -0.0174
#>    120        5.2081             nan     0.1000   -0.0656
#>    140        4.6673             nan     0.1000   -0.0214
#>    150        4.4459             nan     0.1000   -0.0155
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       75.7492             nan     0.1000    7.4102
#>      2       69.5362             nan     0.1000    5.6086
#>      3       63.8651             nan     0.1000    5.7476
#>      4       58.9754             nan     0.1000    4.5767
#>      5       54.6450             nan     0.1000    4.2323
#>      6       50.6089             nan     0.1000    3.3921
#>      7       47.4188             nan     0.1000    2.6961
#>      8       44.4796             nan     0.1000    3.1045
#>      9       41.6009             nan     0.1000    2.5797
#>     10       38.7054             nan     0.1000    1.6693
#>     20       23.7970             nan     0.1000    0.7251
#>     40       14.8019             nan     0.1000    0.0838
#>     60       12.6516             nan     0.1000   -0.0017
#>     80       11.5277             nan     0.1000   -0.0043
#>    100       10.9012             nan     0.1000   -0.0351
#>    120       10.4352             nan     0.1000    0.0032
#>    140       10.1349             nan     0.1000   -0.0125
#>    150        9.9622             nan     0.1000   -0.1117
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       72.9991             nan     0.1000    8.9670
#>      2       64.9234             nan     0.1000    7.3709
#>      3       58.2540             nan     0.1000    6.3747
#>      4       51.3230             nan     0.1000    5.5816
#>      5       45.7691             nan     0.1000    5.5955
#>      6       41.5788             nan     0.1000    3.9459
#>      7       38.0770             nan     0.1000    3.0675
#>      8       34.4352             nan     0.1000    2.9593
#>      9       31.8519             nan     0.1000    2.5459
#>     10       29.5651             nan     0.1000    2.2197
#>     20       17.3182             nan     0.1000    0.6937
#>     40       12.2656             nan     0.1000    0.0466
#>     60       10.3492             nan     0.1000   -0.0733
#>     80        9.0895             nan     0.1000    0.1033
#>    100        8.1133             nan     0.1000    0.0232
#>    120        7.3365             nan     0.1000   -0.0570
#>    140        6.8603             nan     0.1000   -0.0589
#>    150        6.6114             nan     0.1000    0.0053
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       70.9921             nan     0.1000   13.4417
#>      2       61.9951             nan     0.1000    6.6366
#>      3       55.4162             nan     0.1000    6.4825
#>      4       48.5613             nan     0.1000    5.4324
#>      5       43.3730             nan     0.1000    5.3471
#>      6       39.0133             nan     0.1000    4.6963
#>      7       35.0712             nan     0.1000    3.2889
#>      8       31.6762             nan     0.1000    3.5873
#>      9       28.8756             nan     0.1000    2.5564
#>     10       26.4250             nan     0.1000    2.1134
#>     20       14.1161             nan     0.1000    0.3893
#>     40        8.8484             nan     0.1000    0.0388
#>     60        7.2099             nan     0.1000   -0.0085
#>     80        6.3385             nan     0.1000   -0.0037
#>    100        5.6509             nan     0.1000   -0.0548
#>    120        5.0596             nan     0.1000   -0.1034
#>    140        4.6165             nan     0.1000   -0.0355
#>    150        4.3701             nan     0.1000   -0.0054
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       73.7415             nan     0.1000    9.9943
#>      2       64.1152             nan     0.1000    8.1282
#>      3       56.2710             nan     0.1000    7.0320
#>      4       50.0664             nan     0.1000    5.7697
#>      5       44.4789             nan     0.1000    5.1486
#>      6       39.6332             nan     0.1000    4.7685
#>      7       35.9211             nan     0.1000    3.5583
#>      8       32.3251             nan     0.1000    3.7663
#>      9       29.5348             nan     0.1000    2.4861
#>     10       27.1075             nan     0.1000    1.5801
#>     20       15.0626             nan     0.1000    0.3791
#>     40        9.5531             nan     0.1000    0.0239
#>     60        7.8993             nan     0.1000   -0.0318
#>     80        6.7816             nan     0.1000   -0.0424
#>    100        6.0136             nan     0.1000   -0.0362
#>    120        5.3205             nan     0.1000   -0.0291
#>    140        4.8346             nan     0.1000   -0.0399
#>    150        4.6396             nan     0.1000   -0.0038
#> 

boston.inst <- Boston[370,1:13]
ciu <- ciu.new(gbm, medv~., Boston)
cat(ciu.textual(ciu, boston.inst,use.text.effects = TRUE))

#> The value of output 'medv' for instance '370' is 49.187, which is very good (CU=0.982).
#> Feature 'lstat' is important (CI=0.447) and value '3.73' is very good (CU=1).
#> Feature 'dis' is slightly important (CI=0.382) and value '1.3567' is very good (CU=1).
#> Feature 'rm' is slightly important (CI=0.216) and value '6.683' is bad (CU=0.27).
#> Feature 'crim' is not important (CI=0.116) and value '5.66998' is very good (CU=1).
#> Feature 'nox' is not important (CI=0.107) and value '0.631' is very good (CU=1).
#> Feature 'black' is not important (CI=0.083) and value '375.33' is very good (CU=0.871).
#> Feature 'ptratio' is not important (CI=0.064) and value '20.2' is very bad (CU=0.151).
#> Feature 'rad' is not important (CI=0.041) and value '24' is very good (CU=1).
#> Feature 'age' is not important (CI=0.039) and value '96.8' is bad (CU=0.349).
#> Feature 'chas' is not important (CI=0.033) and value '1' is very good (CU=1).
#> Feature 'tax' is not important (CI=0.027) and value '666' is very bad (CU=0).
#> Feature 'indus' is not important (CI=0.02) and value '18.1' is average (CU=0.477).
#> Feature 'zn' is not important (CI=0) and value '0' is very bad (CU=0).

# Customized limits for CI.
cat(ciu.textual(ciu, boston.inst,use.text.effects = TRUE,
  CI.voc = data.frame(limits=c(0.05,0.1,0.3,0.5,1.0),
texts=c("not important","little important", "important","very important",
  "extremely important"))))

#> The value of output 'medv' for instance '370' is 49.187, which is very good (CU=0.982).
#> Feature 'lstat' is very important (CI=0.447) and value '3.73' is very good (CU=1).
#> Feature 'dis' is very important (CI=0.382) and value '1.3567' is very good (CU=1).
#> Feature 'rm' is important (CI=0.21) and value '6.683' is bad (CU=0.246).
#> Feature 'crim' is important (CI=0.116) and value '5.66998' is very good (CU=1).
#> Feature 'nox' is important (CI=0.107) and value '0.631' is very good (CU=1).
#> Feature 'black' is little important (CI=0.083) and value '375.33' is very good (CU=0.871).
#> Feature 'ptratio' is little important (CI=0.064) and value '20.2' is very bad (CU=0.151).
#> Feature 'age' is not important (CI=0.043) and value '96.8' is average (CU=0.4).
#> Feature 'rad' is not important (CI=0.041) and value '24' is very good (CU=1).
#> Feature 'chas' is not important (CI=0.033) and value '1' is very good (CU=1).
#> Feature 'tax' is not important (CI=0.027) and value '666' is very bad (CU=0).
#> Feature 'indus' is not important (CI=0.02) and value '18.1' is average (CU=0.477).
#> Feature 'zn' is not important (CI=0) and value '0' is very bad (CU=0).

# Intermediate concepts
social<-c(1,11,13); usage_type<-c(2,3); chas<-c(4); air_quality<-c(5)
housing<-c(6,7); transport<-c(8,9); blacks<-c(12); tax<-c(10)
Boston.voc <- list("SOCIAL"=social, "LAND USAGE"=usage_type, "Charles R. dummy"=chas,
"Air quality (Nox)"=air_quality, "HOUSING"=housing, "TRANSPORT"=transport,
"Prop. of black people"=blacks, "Tax"=tax)
ciu <- ciu.new(gbm, medv~., Boston, vocabulary = Boston.voc)

# We use `meta.explain` here to avoid differences due to sampling.
meta.top <- ciu$meta.explain(boston.inst, concepts.to.explain=names(Boston.voc))
cat(ciu.textual(ciu, boston.inst, use.text.effects = TRUE, ciu.meta = meta.top))

#> The value of output 'medv' for instance '370' is 49.187, which is very good (CU=0.982).
#> Feature 'SOCIAL' is very important (CI=0.642) and value is very good (CU=0.916).
#> Feature 'TRANSPORT' is important (CI=0.423) and value is very good (CU=1).
#> Feature 'HOUSING' is slightly important (CI=0.248) and value is bad (CU=0.262).
#> Feature 'Air quality (Nox)' is not important (CI=0.107) and value '0.631' is very good (CU=1).
#> Feature 'Prop. of black people' is not important (CI=0.077) and value '375.33' is very good (CU=0.938).
#> Feature 'Charles R. dummy' is not important (CI=0.033) and value '1' is very good (CU=1).
#> Feature 'Tax' is not important (CI=0.027) and value '666' is very bad (CU=0).
#> Feature 'LAND USAGE' is not important (CI=0.02) and value is average (CU=0.477).

# Explain intermediate concept utility, using input features (could also
# be using other intermediate concepts).
cat(ciu.textual(ciu, boston.inst, use.text.effects = TRUE, ind.inputs = Boston.voc$SOCIAL,
  target.concept = "SOCIAL", target.ciu = meta.top$ciuvals[["SOCIAL"]]))

#> The value of intermediate concept 'SOCIAL' for output 'medv', with instance '370' is very good (CU=0.916).
#> Feature 'lstat' is very important (CI=0.696) and value '3.73' is very good (CU=1).
#> Feature 'crim' is not important (CI=0.181) and value '5.66998' is very good (CU=1).
#> Feature 'ptratio' is not important (CI=0.099) and value '20.2' is very bad (CU=0.151).

cat(ciu.textual(ciu, boston.inst, use.text.effects = TRUE, ind.inputs = Boston.voc$HOUSING,
  target.concept = "HOUSING", target.ciu = meta.top$ciuvals[["HOUSING"]]))

#> The value of intermediate concept 'HOUSING' for output 'medv', with instance '370' is bad (CU=0.262).
#> Feature 'rm' is extremely important (CI=0.869) and value '6.683' is bad (CU=0.27).
#> Feature 'age' is not important (CI=0.159) and value '96.8' is bad (CU=0.349).