ggplot2 "col" plot CIU explanation for specific instance

Create a barplot showing CI as the length of the bar and CU on color scale from red to green, via yellow, for the given inputs and the given output. First get a CIU object by calling ciu.new as e.g. ciu <- ciu.new(...), then call ciu.res <- ciu$barplot.ciu(...). "Usage" section is in "Details" section because Roxygen etc. don't support documentation of functions within functions.

Arguments

instance

Instance to explain. See explain.

ind.inputs

vector of indices for the inputs to be included in the plot. If NULL then all inputs will be included.

output.names

Vector with names of outputs to include. If NULL (default), then include all.

in.min.max.limits

See explain.

n.samples

See explain.

neutral.CU

Indicates when the Contextual Utility is considered to be "negative". The default value of 0.5 seems quite logical for most cases.

show.input.values

Include input values after input labels or not.

concepts.to.explain

List of concepts to use in the plot, as defined by vocabulary provided as argument to ciu.new. If ind.inputs=NULL, then use concepts.to.explain instead. If both are NULL, then use all inputs.

target.concept

See explain.

target.ciu

See explain.

low.color

Color to use for CU=0. Default is red.

mid.color

Color to use for CU=neutral.CU. Default is yellow.

high.color

Color to use for CU=1. Default is darkgreen.

sort

NOT USED FOR THE MOMENT! Features are in the same order for all facets, sorted by mean importance over all facets, which feels like a decent behaviour. NULL, "CI" or "CU". No sorting by default, other options are sorting by CI or CU.

decreasing

NOT USED FOR THE MOMENT. Set to TRUE for decreasing sort.

main

Replace default main title of plot.

Details

First get a CIU object by calling ciu.new as e.g. ciu <- ciu.new(...), then call ciu.res <- ciu$ggplot.col.ciu(...). "Usage" section is here in "Details" section because Roxygen etc. don't support documentation of functions within functions.

Usage

ggplot.col.ciu(
  instance,
  ind.inputs=NULL,
  output.names=NULL,
  in.min.max.limits=NULL,
  n.samples=100,
  neutral.CU=0.5,
  show.input.values=TRUE,
  concepts.to.explain=NULL,
  target.concept=NULL,
  target.ciu=NULL,
  low.color="red",
  mid.color="yellow",
  high.color="darkgreen",
  sort=NULL,
  decreasing=FALSE,
  main=NULL)

Value

Created ggplot object.

See also

Author

Kary Främling