Plot output value as a function of one input for a specific instance

Plot how the value of one output changes as a function of one input, as a line chart. The current input/output values are indicated by a red dot. The values of all other inputs are the ones given by the instance parameter. This method is not specific for CIU but it allows to study the behaviour of the underlying "black-box model". It also makes it easy to understand how CI and CU values have been calculated.

Arguments

instance

Instance to explain. See explain.

ind.input

Index of the input to plot

ind.output

Index of the output to plot

in.min.max.limits

See explain.

n.points

The number of points to use on X-axis for plotting.

main, xlab, ylab, ylim, ...

Usual plot parameters, possible to override the default ones provided here if needed.

Details

First get a CIU object by calling ciu.new as e.g. ciu <- ciu.new(...), then call ciu.res <- ciu$plot.ciu(...). "Usage" section is here in "Details" section because Roxygen etc. don't support documentation of functions within functions. Usage

plot.ciu(
  instance,
  ind.input,
  ind.output,
  in.min.max.limits=NULL,
  n.points=40,
  main=NULL,
  xlab=NULL,
  ylab=NULL,
  ylim=NULL,
  ...)

Value

"void", i.e. whatever happens to be result of last instruction.

Author

Kary Främling