Give textual CIU explanation

Provide textual CIU explanations as those used in Kary Främling's PhD thesis.

ciu.textual(
  ciu,
  instance = NULL,
  ind.inputs = NULL,
  ind.output = 1,
  in.min.max.limits = NULL,
  n.samples = 100,
  neutral.CU = 0.5,
  show.input.values = TRUE,
  concepts.to.explain = NULL,
  target.concept = NULL,
  target.ciu = NULL,
  ciu.meta = NULL,
  sort = "CI",
  n.features = NULL,
  use.text.effects = FALSE,
  CI.voc = data.frame(limits = c(0.2, 0.4, 0.6, 0.8, 1), texts = c("not important",
    "slightly important", "important", "very important", "extremely important")),
  CU.voc = data.frame(limits = c(0.2, 0.4, 0.6, 0.8, 1), texts = c("very bad", "bad",
    "average", "good", "very good"))
)

Arguments

ciu

ciu object as created with ciu function (not to be confused with CIU object as created by ciu.new).

instance

Input values for the instance to explain. Should be a data.frame even though a vector or matrix might work too if input names and other needed metadata can be deduced from the dataset or other parameters given to ciu.new.

ind.inputs

Indices of input features to explain (the set i in CIU formulae)

ind.output

Index of output to be explained.

in.min.max.limits

data.frame or matrix with one row per output and two columns, where the first column indicates the minimal value and the second column the maximal value for that output. ONLY NEEDED HERE IF not given as parameter to ciu.new or if the limits are different for this specific instance than the default ones.

n.samples

How many instances to generate for estimating CI and CU. For inputs of type factor, all possible combinations of input values are generated, so this parameter only influences how many instances are (at least) generated for continuous-valued inputs.

neutral.CU

Indicates when the Contextual Utility is considered to be "negative". The default value of 0.5 seems quite logical for most cases.

show.input.values

Include input values after input labels or not. Default is TRUE.

concepts.to.explain

List of input feature concepts to explain, as defined by vocabulary provided as argument to ciu.new. If ind.inputs=NULL, then use concepts.to.explain instead. If both are NULL, then use all inputs.

target.concept

If provided, then calculate CIU of inputs ind.inputs.to.explain relative to the given concept rather than relative to the actual output(s). ind.inputs.to.explain should normally be a subset (or all) of the inputs that target.concept consists of, even though that not required by the CIU calculation. If a target.ciu is provided, then the target.concept doesn't have to be included in the vocabulary gives as parameter to ciu.new (at least for the moment).

target.ciu

ciu.result object previously calculated for target.concept. If a target.concept is provided but target.ciu=NULL, then target.ciu is estimated by a call to ciu.explain with the n.samples value given as a parameter to this call. It may be useful to provide target.ciu if it should be estimated using some other (typically greater) value for n.samples than the default one, or if it has already been calculated for some reason.

ciu.meta

If given, then use existing ciu.meta.result rather than calling ciu.meta.explain.

sort

NULL, "CI" or "CU".

n.features

Maximal number of features to include in explanation.

use.text.effects

Add bold/italics/colors effects?

CI.voc

Limits and texts to use for CI values.

CU.voc

Limits and texts to use for CU values.

Value

Text string with explanation.

Examples

# Explaining the classification of an Iris instance with lda model.
# We use a versicolor (instance 100).
library(MASS)
test.ind <- 100
iris_test <- iris[test.ind, 1:4]
iris_train <- iris[-test.ind, 1:4]
iris_lab <- iris[[5]][-test.ind]
model <- lda(iris_train, iris_lab)

# Create CIU object
ciu <- ciu.new(model, Species~., iris)

# Give textual explanation. Use 'cat' for getting newlines to work.
cat(ciu.textual(ciu, iris_test, ind.output = 2))
#> The value of output 'versicolor' for instance '100' is 1, which is very good (CU=1).
#> Feature 'Petal.Length' is extremely important (CI=1) and value '4.1' is very good (CU=1).
#> Feature 'Petal.Width' is extremely important (CI=1) and value '1.3' is very good (CU=1).
#> Feature 'Sepal.Length' is not important (CI=0.007) and value '5.7' is very good (CU=0.989).
#> Feature 'Sepal.Width' is not important (CI=0.001) and value '2.8' is very good (CU=0.931).
cat(ciu.textual(ciu, iris_test, ind.output = 2, n.features = 2))
#> The value of output 'versicolor' for instance '100' is 1, which is very good (CU=1).
#> Feature 'Petal.Length' is extremely important (CI=1) and value '4.1' is very good (CU=1).
#> Feature 'Petal.Width' is extremely important (CI=1) and value '1.3' is very good (CU=1).

# Boston housing, GBM model.
library(caret)
kfoldcv <- trainControl(method="cv", number=10)
gbm <- train(medv ~ ., Boston, method="gbm", trControl=kfoldcv)
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       78.3553             nan     0.1000    7.0679
#>      2       71.9385             nan     0.1000    6.8015
#>      3       66.8260             nan     0.1000    4.2778
#>      4       62.4584             nan     0.1000    4.5257
#>      5       57.6519             nan     0.1000    3.9490
#>      6       53.5762             nan     0.1000    4.0622
#>      7       50.1194             nan     0.1000    3.0290
#>      8       46.4856             nan     0.1000    2.7993
#>      9       43.2240             nan     0.1000    2.8787
#>     10       40.5302             nan     0.1000    2.5802
#>     20       25.6605             nan     0.1000    0.8864
#>     40       16.7366             nan     0.1000    0.1705
#>     60       13.6860             nan     0.1000   -0.0020
#>     80       12.1916             nan     0.1000    0.0368
#>    100       11.1929             nan     0.1000   -0.0205
#>    120       10.7108             nan     0.1000   -0.0038
#>    140       10.3046             nan     0.1000    0.0187
#>    150       10.1399             nan     0.1000   -0.0266
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       74.6923             nan     0.1000    9.4095
#>      2       66.2790             nan     0.1000    6.9650
#>      3       59.4605             nan     0.1000    7.2597
#>      4       53.1384             nan     0.1000    6.7393
#>      5       47.5877             nan     0.1000    4.6430
#>      6       43.0270             nan     0.1000    3.6956
#>      7       39.0005             nan     0.1000    3.8046
#>      8       35.6987             nan     0.1000    2.6134
#>      9       32.9548             nan     0.1000    2.3282
#>     10       30.4891             nan     0.1000    2.2755
#>     20       18.4021             nan     0.1000    0.4215
#>     40       12.1911             nan     0.1000    0.0249
#>     60       10.1484             nan     0.1000   -0.0356
#>     80        9.0208             nan     0.1000   -0.1754
#>    100        8.2296             nan     0.1000   -0.0411
#>    120        7.5650             nan     0.1000   -0.0041
#>    140        7.0623             nan     0.1000   -0.0635
#>    150        6.7927             nan     0.1000    0.0061
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       73.4379             nan     0.1000   10.3356
#>      2       63.5803             nan     0.1000    8.0995
#>      3       55.7362             nan     0.1000    6.6707
#>      4       49.6750             nan     0.1000    5.8394
#>      5       43.9564             nan     0.1000    5.7512
#>      6       39.3788             nan     0.1000    4.4116
#>      7       35.4454             nan     0.1000    3.0495
#>      8       31.9836             nan     0.1000    2.4603
#>      9       29.0387             nan     0.1000    2.7599
#>     10       26.3064             nan     0.1000    2.1503
#>     20       14.4341             nan     0.1000    0.3877
#>     40        8.7654             nan     0.1000    0.0807
#>     60        7.1621             nan     0.1000    0.0167
#>     80        6.2280             nan     0.1000   -0.0206
#>    100        5.6063             nan     0.1000   -0.0091
#>    120        5.1205             nan     0.1000   -0.0088
#>    140        4.6054             nan     0.1000   -0.0302
#>    150        4.4259             nan     0.1000   -0.0265
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       76.0615             nan     0.1000    6.0268
#>      2       69.8394             nan     0.1000    5.8948
#>      3       64.6439             nan     0.1000    4.6761
#>      4       60.1727             nan     0.1000    4.4242
#>      5       55.5343             nan     0.1000    4.0851
#>      6       51.3024             nan     0.1000    3.0597
#>      7       47.8339             nan     0.1000    3.8949
#>      8       45.2526             nan     0.1000    1.9971
#>      9       42.6343             nan     0.1000    2.0127
#>     10       40.0624             nan     0.1000    2.7514
#>     20       24.2048             nan     0.1000    0.6492
#>     40       15.5930             nan     0.1000   -0.1561
#>     60       13.1375             nan     0.1000    0.0567
#>     80       11.7734             nan     0.1000   -0.0094
#>    100       10.9645             nan     0.1000   -0.0154
#>    120       10.3645             nan     0.1000   -0.0593
#>    140        9.9826             nan     0.1000    0.0029
#>    150        9.8403             nan     0.1000   -0.0751
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       73.5946             nan     0.1000    9.1857
#>      2       64.2093             nan     0.1000    9.0061
#>      3       56.9555             nan     0.1000    7.0134
#>      4       51.1729             nan     0.1000    4.9754
#>      5       45.6779             nan     0.1000    4.4291
#>      6       41.0642             nan     0.1000    3.6577
#>      7       37.5319             nan     0.1000    2.7374
#>      8       34.3908             nan     0.1000    2.9055
#>      9       31.5780             nan     0.1000    1.9422
#>     10       29.0991             nan     0.1000    1.8676
#>     20       16.7319             nan     0.1000    0.5788
#>     40       11.1103             nan     0.1000   -0.0211
#>     60        9.1806             nan     0.1000    0.0068
#>     80        8.2254             nan     0.1000   -0.0268
#>    100        7.3500             nan     0.1000   -0.0230
#>    120        6.6881             nan     0.1000   -0.0210
#>    140        6.2678             nan     0.1000   -0.0417
#>    150        6.0991             nan     0.1000   -0.0846
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       71.8699             nan     0.1000   11.2176
#>      2       62.4685             nan     0.1000    8.8899
#>      3       54.8604             nan     0.1000    6.2612
#>      4       48.5298             nan     0.1000    6.1190
#>      5       43.0929             nan     0.1000    6.2679
#>      6       38.1848             nan     0.1000    4.7041
#>      7       34.4558             nan     0.1000    3.3028
#>      8       31.3115             nan     0.1000    2.7639
#>      9       28.5733             nan     0.1000    2.2043
#>     10       26.2152             nan     0.1000    2.3866
#>     20       14.3189             nan     0.1000    0.3074
#>     40        9.1059             nan     0.1000   -0.0305
#>     60        7.3002             nan     0.1000   -0.0002
#>     80        6.3170             nan     0.1000    0.0261
#>    100        5.6294             nan     0.1000   -0.0487
#>    120        4.9877             nan     0.1000   -0.0499
#>    140        4.5299             nan     0.1000   -0.0364
#>    150        4.3537             nan     0.1000   -0.0631
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       75.7625             nan     0.1000    6.5824
#>      2       69.8683             nan     0.1000    6.1760
#>      3       64.9867             nan     0.1000    5.1694
#>      4       60.0212             nan     0.1000    5.1730
#>      5       55.9137             nan     0.1000    3.9597
#>      6       52.2467             nan     0.1000    3.4249
#>      7       48.8973             nan     0.1000    2.7863
#>      8       46.0269             nan     0.1000    2.4455
#>      9       42.9569             nan     0.1000    2.5536
#>     10       40.5561             nan     0.1000    2.0477
#>     20       25.5246             nan     0.1000    0.5767
#>     40       16.4672             nan     0.1000    0.1125
#>     60       13.8158             nan     0.1000    0.0079
#>     80       12.4329             nan     0.1000   -0.0378
#>    100       11.6889             nan     0.1000    0.0036
#>    120       11.0661             nan     0.1000   -0.0143
#>    140       10.7160             nan     0.1000   -0.0231
#>    150       10.5246             nan     0.1000   -0.0370
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       73.4347             nan     0.1000    7.2333
#>      2       65.7966             nan     0.1000    8.1133
#>      3       58.8093             nan     0.1000    5.3583
#>      4       52.5362             nan     0.1000    5.4932
#>      5       47.0236             nan     0.1000    5.3185
#>      6       42.6032             nan     0.1000    4.0847
#>      7       38.4516             nan     0.1000    3.8946
#>      8       35.2209             nan     0.1000    2.4346
#>      9       32.7206             nan     0.1000    2.5484
#>     10       30.4398             nan     0.1000    1.9844
#>     20       18.3219             nan     0.1000    0.5352
#>     40       12.0648             nan     0.1000   -0.2333
#>     60        9.9154             nan     0.1000    0.0629
#>     80        8.8187             nan     0.1000    0.0356
#>    100        7.8066             nan     0.1000   -0.0499
#>    120        7.1483             nan     0.1000   -0.0008
#>    140        6.8124             nan     0.1000   -0.0259
#>    150        6.6066             nan     0.1000   -0.0484
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       71.4552             nan     0.1000   11.2063
#>      2       63.6114             nan     0.1000    7.6839
#>      3       55.8862             nan     0.1000    7.2312
#>      4       49.7358             nan     0.1000    5.5786
#>      5       44.6420             nan     0.1000    4.9598
#>      6       39.8138             nan     0.1000    4.8409
#>      7       35.9382             nan     0.1000    3.0447
#>      8       32.7022             nan     0.1000    2.7211
#>      9       30.0848             nan     0.1000    2.6056
#>     10       27.7162             nan     0.1000    1.7827
#>     20       15.7474             nan     0.1000    0.3824
#>     40       10.3397             nan     0.1000    0.0503
#>     60        8.3949             nan     0.1000    0.0814
#>     80        7.0706             nan     0.1000   -0.0542
#>    100        6.0506             nan     0.1000   -0.0121
#>    120        5.4412             nan     0.1000   -0.0383
#>    140        5.0080             nan     0.1000   -0.0848
#>    150        4.7368             nan     0.1000   -0.0334
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       79.3169             nan     0.1000    7.1680
#>      2       73.2386             nan     0.1000    5.2351
#>      3       66.1257             nan     0.1000    4.9287
#>      4       61.2451             nan     0.1000    4.4879
#>      5       57.1254             nan     0.1000    4.0334
#>      6       52.9796             nan     0.1000    4.5098
#>      7       49.7005             nan     0.1000    2.5286
#>      8       46.3574             nan     0.1000    2.5148
#>      9       43.5543             nan     0.1000    2.2788
#>     10       40.9323             nan     0.1000    2.7060
#>     20       25.8218             nan     0.1000    0.5297
#>     40       16.8598             nan     0.1000    0.1533
#>     60       14.1874             nan     0.1000   -0.1885
#>     80       12.8855             nan     0.1000    0.0027
#>    100       12.1317             nan     0.1000   -0.1672
#>    120       11.5389             nan     0.1000   -0.0699
#>    140       10.9137             nan     0.1000   -0.0102
#>    150       10.6855             nan     0.1000   -0.0721
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       76.6838             nan     0.1000   11.4692
#>      2       69.3883             nan     0.1000    7.6412
#>      3       61.6957             nan     0.1000    6.9343
#>      4       55.2718             nan     0.1000    5.9045
#>      5       49.9783             nan     0.1000    5.4273
#>      6       45.1316             nan     0.1000    4.0962
#>      7       41.3303             nan     0.1000    3.5726
#>      8       37.8017             nan     0.1000    2.2711
#>      9       34.7064             nan     0.1000    2.3184
#>     10       31.9668             nan     0.1000    1.7941
#>     20       18.8183             nan     0.1000    0.4331
#>     40       12.6134             nan     0.1000    0.0849
#>     60       10.5149             nan     0.1000   -0.0692
#>     80        9.2890             nan     0.1000   -0.0475
#>    100        8.3215             nan     0.1000   -0.0075
#>    120        7.5798             nan     0.1000   -0.0141
#>    140        6.9621             nan     0.1000   -0.0671
#>    150        6.7430             nan     0.1000   -0.0153
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       75.5637             nan     0.1000   10.6884
#>      2       65.5195             nan     0.1000    8.8326
#>      3       56.9188             nan     0.1000    8.8839
#>      4       50.2184             nan     0.1000    6.2916
#>      5       44.9185             nan     0.1000    4.6751
#>      6       39.9172             nan     0.1000    4.7753
#>      7       36.3137             nan     0.1000    3.1596
#>      8       33.0034             nan     0.1000    2.9855
#>      9       29.9466             nan     0.1000    2.4298
#>     10       27.6281             nan     0.1000    2.1369
#>     20       15.8088             nan     0.1000    0.0415
#>     40        9.6973             nan     0.1000    0.0358
#>     60        7.8008             nan     0.1000   -0.0464
#>     80        6.7191             nan     0.1000   -0.0356
#>    100        5.9541             nan     0.1000   -0.0513
#>    120        5.2945             nan     0.1000   -0.0391
#>    140        4.7779             nan     0.1000   -0.0399
#>    150        4.5290             nan     0.1000   -0.0076
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       79.6618             nan     0.1000    7.4213
#>      2       73.2324             nan     0.1000    6.8411
#>      3       67.4767             nan     0.1000    5.6002
#>      4       62.4106             nan     0.1000    4.8971
#>      5       57.8117             nan     0.1000    4.6352
#>      6       53.6685             nan     0.1000    3.1054
#>      7       50.2983             nan     0.1000    3.6634
#>      8       46.9772             nan     0.1000    3.4329
#>      9       44.3396             nan     0.1000    2.2368
#>     10       41.8497             nan     0.1000    2.6047
#>     20       26.3159             nan     0.1000    0.9086
#>     40       17.5714             nan     0.1000    0.0008
#>     60       14.3752             nan     0.1000    0.0154
#>     80       12.7861             nan     0.1000    0.0238
#>    100       11.9160             nan     0.1000    0.0308
#>    120       11.2816             nan     0.1000   -0.0133
#>    140       10.8119             nan     0.1000    0.0192
#>    150       10.6266             nan     0.1000   -0.0272
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       76.6551             nan     0.1000    8.6940
#>      2       68.4576             nan     0.1000    8.1906
#>      3       61.1683             nan     0.1000    7.1286
#>      4       54.6586             nan     0.1000    5.1685
#>      5       48.8526             nan     0.1000    5.4502
#>      6       44.2997             nan     0.1000    3.7423
#>      7       40.0908             nan     0.1000    3.3014
#>      8       37.0001             nan     0.1000    3.0772
#>      9       34.0615             nan     0.1000    2.4425
#>     10       31.5773             nan     0.1000    2.0747
#>     20       18.9148             nan     0.1000    0.7231
#>     40       12.0888             nan     0.1000   -0.0453
#>     60        9.9497             nan     0.1000   -0.0221
#>     80        8.8292             nan     0.1000   -0.0368
#>    100        7.9724             nan     0.1000   -0.1161
#>    120        7.4277             nan     0.1000   -0.0754
#>    140        6.8779             nan     0.1000   -0.0560
#>    150        6.6479             nan     0.1000   -0.0998
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       76.0695             nan     0.1000   11.1969
#>      2       65.8340             nan     0.1000    7.3695
#>      3       59.2526             nan     0.1000    7.2014
#>      4       52.7956             nan     0.1000    6.8461
#>      5       46.7443             nan     0.1000    6.0340
#>      6       41.5525             nan     0.1000    5.0642
#>      7       37.4298             nan     0.1000    4.0635
#>      8       34.1839             nan     0.1000    2.7936
#>      9       31.1715             nan     0.1000    2.5798
#>     10       28.6871             nan     0.1000    2.3534
#>     20       15.9171             nan     0.1000    0.3865
#>     40        9.7806             nan     0.1000   -0.0623
#>     60        8.0435             nan     0.1000    0.0019
#>     80        6.9040             nan     0.1000   -0.0566
#>    100        5.9436             nan     0.1000   -0.0290
#>    120        5.2445             nan     0.1000   -0.0080
#>    140        4.8288             nan     0.1000   -0.0650
#>    150        4.6399             nan     0.1000   -0.0250
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       75.1757             nan     0.1000    5.4262
#>      2       68.5510             nan     0.1000    6.3972
#>      3       62.6764             nan     0.1000    4.5340
#>      4       57.6049             nan     0.1000    4.5640
#>      5       53.2824             nan     0.1000    4.1536
#>      6       49.4824             nan     0.1000    3.0501
#>      7       46.4331             nan     0.1000    2.4064
#>      8       43.1393             nan     0.1000    3.1151
#>      9       40.2981             nan     0.1000    2.2694
#>     10       37.8182             nan     0.1000    2.4685
#>     20       23.5047             nan     0.1000    0.7099
#>     40       15.2273             nan     0.1000    0.0775
#>     60       13.0322             nan     0.1000   -0.1264
#>     80       11.8016             nan     0.1000    0.0501
#>    100       11.0568             nan     0.1000   -0.0485
#>    120       10.6190             nan     0.1000   -0.0226
#>    140       10.2199             nan     0.1000   -0.1275
#>    150        9.9890             nan     0.1000   -0.0591
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       73.1092             nan     0.1000   10.3637
#>      2       65.0285             nan     0.1000    8.0354
#>      3       58.0597             nan     0.1000    8.1663
#>      4       51.0852             nan     0.1000    6.0292
#>      5       45.8006             nan     0.1000    4.9421
#>      6       41.5433             nan     0.1000    3.6584
#>      7       37.8380             nan     0.1000    3.8490
#>      8       34.2208             nan     0.1000    2.9494
#>      9       31.2964             nan     0.1000    2.8128
#>     10       28.8656             nan     0.1000    2.1513
#>     20       16.2651             nan     0.1000    0.3747
#>     40       10.8648             nan     0.1000    0.0502
#>     60        9.0081             nan     0.1000   -0.0107
#>     80        8.2456             nan     0.1000   -0.0635
#>    100        7.5742             nan     0.1000   -0.0549
#>    120        6.9366             nan     0.1000   -0.0546
#>    140        6.4213             nan     0.1000   -0.0319
#>    150        6.1609             nan     0.1000   -0.0676
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       71.2923             nan     0.1000   12.0713
#>      2       62.7069             nan     0.1000    8.2159
#>      3       55.3811             nan     0.1000    8.2913
#>      4       48.7767             nan     0.1000    5.6027
#>      5       43.7313             nan     0.1000    4.7458
#>      6       39.1197             nan     0.1000    3.6196
#>      7       35.5526             nan     0.1000    3.8056
#>      8       32.0906             nan     0.1000    2.7373
#>      9       29.2909             nan     0.1000    2.7599
#>     10       27.0256             nan     0.1000    1.7869
#>     20       14.8291             nan     0.1000    0.0303
#>     40        9.5117             nan     0.1000    0.0314
#>     60        7.8230             nan     0.1000   -0.0978
#>     80        6.7035             nan     0.1000   -0.0545
#>    100        5.9646             nan     0.1000   -0.0190
#>    120        5.4382             nan     0.1000   -0.0450
#>    140        5.0244             nan     0.1000   -0.1344
#>    150        4.8129             nan     0.1000   -0.0662
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       76.4858             nan     0.1000    6.9782
#>      2       70.6749             nan     0.1000    6.8207
#>      3       65.0766             nan     0.1000    5.3767
#>      4       59.9483             nan     0.1000    4.3221
#>      5       55.4110             nan     0.1000    3.9261
#>      6       51.4140             nan     0.1000    3.9962
#>      7       48.0122             nan     0.1000    3.2010
#>      8       44.8311             nan     0.1000    2.4837
#>      9       42.4725             nan     0.1000    2.2332
#>     10       39.8279             nan     0.1000    2.5310
#>     20       25.5222             nan     0.1000    0.9487
#>     40       16.7512             nan     0.1000    0.1679
#>     60       13.7678             nan     0.1000   -0.0614
#>     80       12.2450             nan     0.1000   -0.0125
#>    100       11.3989             nan     0.1000   -0.0535
#>    120       10.8168             nan     0.1000    0.0148
#>    140       10.3870             nan     0.1000    0.0159
#>    150       10.2529             nan     0.1000   -0.0892
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       74.3942             nan     0.1000    8.5322
#>      2       66.5330             nan     0.1000    7.4050
#>      3       59.8113             nan     0.1000    5.4868
#>      4       53.6442             nan     0.1000    6.4598
#>      5       48.8364             nan     0.1000    4.6230
#>      6       44.8090             nan     0.1000    5.0008
#>      7       40.4764             nan     0.1000    3.9571
#>      8       37.5473             nan     0.1000    3.3067
#>      9       35.0785             nan     0.1000    2.4902
#>     10       32.4860             nan     0.1000    1.9553
#>     20       18.4475             nan     0.1000    0.6125
#>     40       11.8559             nan     0.1000    0.0729
#>     60        9.8590             nan     0.1000   -0.0291
#>     80        8.8070             nan     0.1000   -0.0922
#>    100        8.0046             nan     0.1000   -0.0093
#>    120        7.3227             nan     0.1000   -0.0673
#>    140        6.7701             nan     0.1000   -0.0326
#>    150        6.4941             nan     0.1000   -0.0635
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       72.3948             nan     0.1000   12.2843
#>      2       63.8427             nan     0.1000    7.9593
#>      3       56.2768             nan     0.1000    7.6514
#>      4       49.7239             nan     0.1000    5.6885
#>      5       44.7082             nan     0.1000    4.6512
#>      6       40.3357             nan     0.1000    3.3069
#>      7       36.5693             nan     0.1000    4.1931
#>      8       32.6786             nan     0.1000    3.3512
#>      9       30.0498             nan     0.1000    2.6274
#>     10       27.6289             nan     0.1000    2.6182
#>     20       15.4295             nan     0.1000    0.5380
#>     40        9.8184             nan     0.1000    0.0768
#>     60        7.7880             nan     0.1000   -0.0190
#>     80        6.8544             nan     0.1000   -0.0550
#>    100        6.0436             nan     0.1000   -0.0723
#>    120        5.4591             nan     0.1000   -0.0169
#>    140        4.8989             nan     0.1000   -0.0516
#>    150        4.6627             nan     0.1000   -0.0298
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       78.5879             nan     0.1000    6.6094
#>      2       71.8880             nan     0.1000    5.7863
#>      3       65.5290             nan     0.1000    5.9839
#>      4       60.9307             nan     0.1000    4.8747
#>      5       56.7427             nan     0.1000    3.5569
#>      6       52.9696             nan     0.1000    4.2732
#>      7       49.5649             nan     0.1000    3.5562
#>      8       46.8402             nan     0.1000    2.8776
#>      9       44.0113             nan     0.1000    1.8740
#>     10       41.6104             nan     0.1000    2.4570
#>     20       25.7271             nan     0.1000    0.7278
#>     40       17.0028             nan     0.1000    0.1544
#>     60       14.1834             nan     0.1000    0.0445
#>     80       12.5790             nan     0.1000    0.0092
#>    100       11.4927             nan     0.1000   -0.0051
#>    120       10.8930             nan     0.1000   -0.0716
#>    140       10.3435             nan     0.1000   -0.0323
#>    150       10.1405             nan     0.1000   -0.0346
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       75.2918             nan     0.1000   10.5638
#>      2       66.2662             nan     0.1000    9.9381
#>      3       58.3744             nan     0.1000    7.1255
#>      4       52.4120             nan     0.1000    5.1918
#>      5       47.7825             nan     0.1000    4.0530
#>      6       43.5618             nan     0.1000    4.2292
#>      7       39.5942             nan     0.1000    3.6163
#>      8       36.0353             nan     0.1000    2.3627
#>      9       33.3733             nan     0.1000    2.1499
#>     10       30.6108             nan     0.1000    2.4876
#>     20       17.8023             nan     0.1000    0.3143
#>     40       11.6603             nan     0.1000   -0.1018
#>     60        9.8347             nan     0.1000   -0.0854
#>     80        8.7326             nan     0.1000   -0.0789
#>    100        7.9638             nan     0.1000   -0.0238
#>    120        7.3567             nan     0.1000   -0.0251
#>    140        6.8165             nan     0.1000   -0.0617
#>    150        6.6219             nan     0.1000   -0.0201
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       75.3454             nan     0.1000   11.9246
#>      2       65.7944             nan     0.1000    9.4345
#>      3       58.2744             nan     0.1000    6.8293
#>      4       51.2956             nan     0.1000    6.1996
#>      5       45.6956             nan     0.1000    4.8794
#>      6       41.1488             nan     0.1000    4.9876
#>      7       36.9122             nan     0.1000    4.3108
#>      8       33.2689             nan     0.1000    3.4905
#>      9       30.5633             nan     0.1000    2.1255
#>     10       27.8416             nan     0.1000    2.1932
#>     20       15.0068             nan     0.1000    0.4797
#>     40        9.8609             nan     0.1000    0.0531
#>     60        7.8263             nan     0.1000   -0.0431
#>     80        6.8645             nan     0.1000   -0.0600
#>    100        6.0677             nan     0.1000   -0.0347
#>    120        5.3877             nan     0.1000   -0.0114
#>    140        4.9793             nan     0.1000   -0.0291
#>    150        4.7781             nan     0.1000   -0.0305
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       74.3307             nan     0.1000    6.9001
#>      2       68.0389             nan     0.1000    5.3274
#>      3       62.3048             nan     0.1000    5.3934
#>      4       57.3801             nan     0.1000    4.2138
#>      5       53.3281             nan     0.1000    3.9824
#>      6       49.0566             nan     0.1000    4.3196
#>      7       45.6508             nan     0.1000    2.8392
#>      8       42.4688             nan     0.1000    3.0263
#>      9       39.5897             nan     0.1000    2.5629
#>     10       36.9069             nan     0.1000    2.4061
#>     20       22.1491             nan     0.1000    0.6185
#>     40       13.8669             nan     0.1000    0.1716
#>     60       11.3847             nan     0.1000    0.0206
#>     80       10.1685             nan     0.1000    0.0111
#>    100        9.5914             nan     0.1000    0.0001
#>    120        9.0166             nan     0.1000   -0.0299
#>    140        8.6782             nan     0.1000   -0.0855
#>    150        8.5010             nan     0.1000   -0.0224
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       70.9226             nan     0.1000    9.4236
#>      2       62.5521             nan     0.1000    9.5604
#>      3       55.8018             nan     0.1000    6.3013
#>      4       49.7404             nan     0.1000    5.8914
#>      5       44.9002             nan     0.1000    4.6073
#>      6       40.5254             nan     0.1000    3.3750
#>      7       37.0437             nan     0.1000    3.6650
#>      8       34.2492             nan     0.1000    2.7001
#>      9       31.0466             nan     0.1000    3.2699
#>     10       28.6180             nan     0.1000    1.9816
#>     20       15.6373             nan     0.1000    0.6378
#>     40       10.3861             nan     0.1000   -0.0137
#>     60        8.8434             nan     0.1000    0.0048
#>     80        7.9018             nan     0.1000   -0.0440
#>    100        7.0212             nan     0.1000   -0.0144
#>    120        6.3837             nan     0.1000   -0.0149
#>    140        6.0588             nan     0.1000   -0.0568
#>    150        5.7932             nan     0.1000   -0.0117
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       69.6339             nan     0.1000   10.9787
#>      2       60.4681             nan     0.1000   10.0575
#>      3       53.8845             nan     0.1000    7.2654
#>      4       47.3416             nan     0.1000    5.9136
#>      5       41.5541             nan     0.1000    5.4310
#>      6       37.2710             nan     0.1000    3.9096
#>      7       33.5271             nan     0.1000    3.5561
#>      8       30.4568             nan     0.1000    2.6727
#>      9       27.4793             nan     0.1000    2.4912
#>     10       25.1678             nan     0.1000    2.2245
#>     20       13.7528             nan     0.1000    0.5129
#>     40        8.6776             nan     0.1000   -0.1314
#>     60        7.2795             nan     0.1000   -0.0594
#>     80        6.3160             nan     0.1000   -0.0239
#>    100        5.5699             nan     0.1000   -0.0190
#>    120        4.9251             nan     0.1000   -0.0409
#>    140        4.3662             nan     0.1000    0.0023
#>    150        4.1673             nan     0.1000   -0.0024
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       77.4888             nan     0.1000    6.7313
#>      2       70.3147             nan     0.1000    6.1592
#>      3       65.3763             nan     0.1000    3.9524
#>      4       60.2504             nan     0.1000    4.4553
#>      5       55.7502             nan     0.1000    3.8313
#>      6       51.7845             nan     0.1000    3.6930
#>      7       48.7155             nan     0.1000    2.7277
#>      8       45.7041             nan     0.1000    2.9543
#>      9       42.9202             nan     0.1000    2.8383
#>     10       40.5411             nan     0.1000    1.8953
#>     20       25.1648             nan     0.1000    0.5704
#>     40       16.3191             nan     0.1000    0.1064
#>     60       13.7287             nan     0.1000    0.0853
#>     80       12.3743             nan     0.1000   -0.0651
#>    100       11.6935             nan     0.1000   -0.1463
#>    120       11.0298             nan     0.1000    0.0024
#>    140       10.5985             nan     0.1000   -0.0297
#>    150       10.3767             nan     0.1000   -0.0242
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       74.6459             nan     0.1000    9.6176
#>      2       65.8590             nan     0.1000    7.1771
#>      3       58.8885             nan     0.1000    7.5301
#>      4       52.6419             nan     0.1000    6.5410
#>      5       47.5970             nan     0.1000    4.9817
#>      6       42.8786             nan     0.1000    4.7718
#>      7       38.6792             nan     0.1000    3.4889
#>      8       35.3894             nan     0.1000    2.9996
#>      9       32.2964             nan     0.1000    2.8341
#>     10       29.6131             nan     0.1000    1.9405
#>     20       18.4803             nan     0.1000    0.3623
#>     40       12.5381             nan     0.1000   -0.0041
#>     60       10.4825             nan     0.1000    0.0393
#>     80        9.0922             nan     0.1000    0.0013
#>    100        8.0045             nan     0.1000   -0.1008
#>    120        7.4453             nan     0.1000   -0.0011
#>    140        6.8606             nan     0.1000    0.0099
#>    150        6.6550             nan     0.1000   -0.0137
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       73.1302             nan     0.1000    9.6527
#>      2       64.0726             nan     0.1000    8.7720
#>      3       57.3839             nan     0.1000    6.5136
#>      4       51.3694             nan     0.1000    5.8413
#>      5       45.6744             nan     0.1000    4.5174
#>      6       40.9079             nan     0.1000    4.9088
#>      7       36.7553             nan     0.1000    3.0176
#>      8       33.5034             nan     0.1000    3.1574
#>      9       30.2296             nan     0.1000    2.5577
#>     10       27.3398             nan     0.1000    2.3587
#>     20       14.4192             nan     0.1000    0.5646
#>     40        9.1053             nan     0.1000    0.0160
#>     60        7.2638             nan     0.1000    0.0042
#>     80        6.3410             nan     0.1000   -0.0281
#>    100        5.5803             nan     0.1000   -0.0048
#>    120        5.0365             nan     0.1000   -0.0530
#>    140        4.5840             nan     0.1000   -0.0506
#>    150        4.3169             nan     0.1000   -0.0050
#> 
#> Iter   TrainDeviance   ValidDeviance   StepSize   Improve
#>      1       72.7964             nan     0.1000   10.8038
#>      2       64.4863             nan     0.1000    7.9343
#>      3       56.9860             nan     0.1000    7.5110
#>      4       50.6113             nan     0.1000    6.7910
#>      5       45.2867             nan     0.1000    5.0561
#>      6       40.7952             nan     0.1000    4.7649
#>      7       37.0030             nan     0.1000    3.2195
#>      8       33.3829             nan     0.1000    3.5819
#>      9       30.5253             nan     0.1000    3.0089
#>     10       27.9322             nan     0.1000    2.2304
#>     20       14.5655             nan     0.1000    0.7089
#>     40        9.0792             nan     0.1000    0.0207
#>     60        7.3526             nan     0.1000   -0.0396
#>     80        6.3149             nan     0.1000   -0.0551
#>    100        5.6010             nan     0.1000   -0.0244
#>    120        5.1064             nan     0.1000   -0.0068
#>    140        4.5980             nan     0.1000   -0.0398
#>    150        4.3586             nan     0.1000   -0.0059
#> 
boston.inst <- Boston[370,1:13]
ciu <- ciu.new(gbm, medv~., Boston)
cat(ciu.textual(ciu, boston.inst,use.text.effects = TRUE))
#> The value of output 'medv' for instance '370' is 50.092, which is very bad (CU=1.002).
#> Feature 'lstat' is important (CI=0.514) and value '3.73' is very good (CU=1).
#> Feature 'dis' is slightly important (CI=0.347) and value '1.3567' is very good (CU=1).
#> Feature 'rm' is slightly important (CI=0.233) and value '6.683' is very bad (CU=0.144).
#> Feature 'crim' is not important (CI=0.131) and value '5.66998' is very good (CU=1).
#> Feature 'nox' is not important (CI=0.129) and value '0.631' is very good (CU=1).
#> Feature 'black' is not important (CI=0.07) and value '375.33' is very good (CU=1).
#> Feature 'age' is not important (CI=0.053) and value '96.8' is very good (CU=0.893).
#> Feature 'ptratio' is not important (CI=0.034) and value '20.2' is average (CU=0.529).
#> Feature 'chas' is not important (CI=0.034) and value '1' is very good (CU=1).
#> Feature 'tax' is not important (CI=0.031) and value '666' is very bad (CU=0).
#> Feature 'rad' is not important (CI=0.021) and value '24' is very good (CU=1).
#> Feature 'indus' is not important (CI=0.012) and value '18.1' is very bad (CU=0).
#> Feature 'zn' is not important (CI=0) and value '0' is very bad (CU=0).
# Customized limits for CI.
cat(ciu.textual(ciu, boston.inst,use.text.effects = TRUE,
  CI.voc = data.frame(limits=c(0.05,0.1,0.3,0.5,1.0),
texts=c("not important","little important", "important","very important",
  "extremely important"))))
#> The value of output 'medv' for instance '370' is 50.092, which is very bad (CU=1.002).
#> Feature 'lstat' is extremely important (CI=0.517) and value '3.73' is very good (CU=0.993).
#> Feature 'dis' is very important (CI=0.347) and value '1.3567' is very good (CU=1).
#> Feature 'rm' is important (CI=0.233) and value '6.683' is very bad (CU=0.144).
#> Feature 'crim' is important (CI=0.131) and value '5.66998' is very good (CU=1).
#> Feature 'nox' is important (CI=0.129) and value '0.631' is very good (CU=1).
#> Feature 'black' is little important (CI=0.07) and value '375.33' is very good (CU=1).
#> Feature 'age' is little important (CI=0.053) and value '96.8' is very good (CU=0.893).
#> Feature 'ptratio' is not important (CI=0.034) and value '20.2' is average (CU=0.529).
#> Feature 'chas' is not important (CI=0.034) and value '1' is very good (CU=1).
#> Feature 'tax' is not important (CI=0.031) and value '666' is very bad (CU=0).
#> Feature 'rad' is not important (CI=0.021) and value '24' is very good (CU=1).
#> Feature 'indus' is not important (CI=0.012) and value '18.1' is very bad (CU=0).
#> Feature 'zn' is not important (CI=0) and value '0' is very bad (CU=0).

# Intermediate concepts
social<-c(1,11,13); usage_type<-c(2,3); chas<-c(4); air_quality<-c(5)
housing<-c(6,7); transport<-c(8,9); blacks<-c(12); tax<-c(10)
Boston.voc <- list("SOCIAL"=social, "LAND USAGE"=usage_type, "Charles R. dummy"=chas,
"Air quality (Nox)"=air_quality, "HOUSING"=housing, "TRANSPORT"=transport,
"Prop. of black people"=blacks, "Tax"=tax)
ciu <- ciu.new(gbm, medv~., Boston, vocabulary = Boston.voc)

# We use `meta.explain` here to avoid differences due to sampling.
meta.top <- ciu$meta.explain(boston.inst, concepts.to.explain=names(Boston.voc))
cat(ciu.textual(ciu, boston.inst, use.text.effects = TRUE, ciu.meta = meta.top))
#> The value of output 'medv' for instance '370' is 50.092, which is very bad (CU=1.002).
#> Feature 'SOCIAL' is very important (CI=0.677) and value is very good (CU=0.984).
#> Feature 'TRANSPORT' is slightly important (CI=0.358) and value is very good (CU=1).
#> Feature 'HOUSING' is slightly important (CI=0.306) and value is bad (CU=0.282).
#> Feature 'Air quality (Nox)' is not important (CI=0.129) and value '0.631' is very good (CU=1).
#> Feature 'Prop. of black people' is not important (CI=0.07) and value '375.33' is very good (CU=1).
#> Feature 'Charles R. dummy' is not important (CI=0.034) and value '1' is very good (CU=1).
#> Feature 'Tax' is not important (CI=0.031) and value '666' is very bad (CU=0).
#> Feature 'LAND USAGE' is not important (CI=0.012) and value is very bad (CU=0).

# Explain intermediate concept utility, using input features (could also
# be using other intermediate concepts).
cat(ciu.textual(ciu, boston.inst, use.text.effects = TRUE, ind.inputs = Boston.voc$SOCIAL,
  target.concept = "SOCIAL", target.ciu = meta.top$ciuvals[["SOCIAL"]]))
#> The value of intermediate concept 'SOCIAL' for output 'medv', with instance '370' is very good (CU=0.984).
#> Feature 'lstat' is very important (CI=0.759) and value '3.73' is very good (CU=1).
#> Feature 'crim' is not important (CI=0.193) and value '5.66998' is very good (CU=1).
#> Feature 'ptratio' is not important (CI=0.051) and value '20.2' is average (CU=0.529).
cat(ciu.textual(ciu, boston.inst, use.text.effects = TRUE, ind.inputs = Boston.voc$HOUSING,
  target.concept = "HOUSING", target.ciu = meta.top$ciuvals[["HOUSING"]]))
#> The value of intermediate concept 'HOUSING' for output 'medv', with instance '370' is bad (CU=0.282).
#> Feature 'rm' is very important (CI=0.76) and value '6.683' is very bad (CU=0.144).
#> Feature 'age' is not important (CI=0.167) and value '96.8' is very good (CU=0.931).